update: code cleanup
