+++ frontend-sb/src/types/roomStatus.ts
@@ -1,0 +1,15 @@
+import sessionUser from './sessionUser';
+
+type roomStatus = {
+  currentDJ: sessionUser | undefined;
+  currentURI: string[] | undefined;
+  connectedUsers: sessionUser[];
+  startsAt: number;
+  endsAt: number;
+  webplayer: {
+    isPlaying: boolean;
+    positionMs: number;
+  };
+};
+
+export default roomStatus;
